pyproject.toml
[tool.poetry]
name = "creating-gif"
version = "0.1.0"
description = ""
authors = ["George Xie <georgexsh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.18.2"
matplotlib = "^3.2.1"
pillow = "^7.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
